Rescaffold MscmpSystTelemetry
The original generation of this project was faulty due to issues in
msbms_build_lib.

With the msbms_build_lib issues resolved, we regenerate the telemetry
project since nothing really had been done in the project.

Note that this is still just a scaffolded initial commit of a new
project skeleton.