Add complete functionality to MscmpSystTelemetry
I think we're done for now.  This adds the key features and capabilities
at this time for the MscmpSystTelemetry module.  It should be noted that
there's probably some clean up, especially around, compliance with
project standards since this module had a fairly strong "vibe coding"
element to both the coding and the documentation; this isn't to say I
let the LLM completely have it's way, but it did most the coding while
I handled the architectural decisions and performed relatively detailed
code review.

Whatever the case, we're pretty close to the mark I think with a
workable version enough so to move on.