Fix invalid DevSupport registry names across proj.
It use to be that our registry names could be simple atoms used directly
as `Registry` names.  This changed at some point and we need to be
explicit as to the expected registry module.  The components in this
commit apparently didn't get the memo so threw errors when they tried to
do things the old way.

Found building out mssub_mcp, fix across the project since DevSupport is
largely just boilerplate.