Fix MscmpSystInstance.Application duplicate update
There is an `upgrade_instances/1` function in the Application API module
which appears to have been copied into an earlier location in the file
at the time when other Instance management functions were added.  At
that time, I apparently didn't remove that code from the original
location.  This commit remedies that issue.