Add MscmpSystLimiter Token Bucket; stubs for other
This commit adds a usefully complete version of the Token Bucket rate
limiting algorithm and the supporting code, including generally useful
code, which makes the token bucket implementation possible and usable.

In addition, we add some stubbing out of the other algorithms which we
will ultimately support, but have not yet fully implemented.  As such
this commit should be seen as purely interim in nature.