Refactor & complete MscmpSystLimiter; algos, tests
This is a pretty expansive commit (arguably could be broken into
smaller pieces).

First we started with algorithm assumptions, meaning what algorithms and
how to implement them, from ExHammer.  Over the course of development
and some back and forth discussion with LLMs. I came to the conclusion
that the leaky bucket algorithm couldn't be properly implemented in a
library like this since we don't do any sort of queue processing ( event
processing what "leaks" from leaky bucket.) Also fixed window and
sliding window were both looking complex to properly implement and it's
not clear that we have a specific use case where a generalized
implementation such as would be in this Component would be useful: so
skip them!

We did design an algorithm which we're calling "semaphore" because of
general concept matches computing related semaphores best.  The basic
idea is there is a pool of "permits" which can be consumed as needed
until exhausted.  The permits, however, can not only be withdrawn from
the pool on resource usage, but also returned to the pool if
appropriate.  For example, database pooling where there are a finite
number of database connections available, but when a connection is
released it can be reused.  Depending on how you use such an algorithm,
you can implement password retry limits (no return of resources) or
other maximum per duration style limitations; there is a time to live
for the semaphore limiter upon which it expires, the next subsequent
request will recreate the limiter with full capacity available.

This commit includes fixes, filling out the API and related
documentation, creating unit, doc, and integration testing.

I think we're "done" with the Component at this point!
(famous last words)