Update msbms_build_config supported dependencies
Remove ExHammer due to maintainers making radical changes for unclear
reasons.  Replaced with MscmpSystLimiter which was added earlier.

Add MscmpSystService.  This is very lightweight behaviour to establish
the general pattern for GenServer based Components.  So far the three
that we have directly defined follow this patterns, so lets enforce that
across the project.