Add MscmpSystService functionality
This adds all of the functionality currently expected of this Component.

Do note that the test suite is a bit off the rails in terms of over-
testing.  We really don't need to test an entire service implementation
which is based on the behaviour we define.  The LLM wrote the tests and,
as they often do, it went a bit off the rails and overdid it.

I think we'll leave for now as it does poke our setup in some ways, but
if it were ever to become an actual problem, an valid early option would
be to simplify the tests down to essentials.