Refactor MscmpSystOptions test suite
In testing upgraded dependencies, it became clear that this old
Component didn't follow our testings standards which evolved after this
Component was developed.

With LLMs a think now, have the LLM update our test suite to match our
current test patterns, which in turn allow it to work with our build
script.