Add MscmpSystTelemetry to MscmpSystHierarchy
I think this is an adequate starting place.  The addition was LLM
generated, and while I think it did a good job, I think it could have
been applied to more functions than it did.  Again, until we really
start making use of the telemetry data, it's not worth worrying too
much about.