Refactor instance_lifecycle_state
In MscmpSystInstance.  The term `instance_state` was getting overloaded
in too many places and a recent change (though I cannot remember where,
perhaps the new limiter Component?) was going to collide on the term
`instance_state`.

To resolve this add "lifecycle" the name of the MscmpSystInstance module
to handle the likely collisions.

In the end I think this was actually resolved by other means, but since
"Instance" is such a generic term, let's keep our specifications for
when to use it and when to make it something more fit for purpose.