Add MscmpSystTelemetry to MscmpSystInstance
This is s reasonable, but definitely not complete start to adding in
telemetry into the MscmpSystInstance component.  This is a largely LLM
add with review.

It was clear during review that this doesn't cover everything that we'd
want covered, but until such time we start using the telemetry for
useful purposes, we're not going to be helped by thoroughness and it
will be better to at least set the precedence of having the telemetry
incorporated in the first place.