Refactor MscmpSystLimiter ProcessUtils unit tests
I think the LLM generated tests were probably adequate for proving the
process_utils.ex code was working as intended, but it did so in an
overboard, extravagant way.  It was testing process dictionary handling
(i.e. `Process`) as much as our own code.

Let's tone all that down a bit and stay focused on our own module in the
unit tests; especially since all of the `MscmpSystService`
implementations are going to be very similar in this regard, and since
`MscmpSystLimiter` is sort of the reference implementation for our
`MscmpSystService` patterns, let's not let too much crufty code from the
LLM get by.