Refactor MscmpSystUtilsData external_name min.
Reducing this minimum to 2 graphemes from 6.  The reason for this is
that we're not expecting uniqueness here and some values which we might
put here could be small.  The LLM test wrote an external name of "Open"
for an EnumItem... not unreasonable given what enums are for... and
that's shorter than 6.  2 is reasonable I think and conceivable ("No",
"On").