Refactor MscmpSystEnums for MscmpSystService/tests
This is fairly significant update to an old part of the system.  In this
commit we bring changes which mack `MscmpSystEnums` compliant with the
standards established by `MscmpSystService` and we use
`MscmpSystSettings` as a model since it bears significant similarity to
`MscmpSystEnums` in overall design; `MscmpSystLimiter` still, however,
remains the reference implementation of `MscmpSystService` however.

In addition we update the `MscmpSystEnums` test suite to include a more
thorough and properly targeted set of unit tests and to create proper
integration tests in the first place.

Note that the integration tests are LLM generated with a significant
amount of guidance before the fact and review after; the unit tests are
LLM generated as well, but with less intervention from me regarding
specifics of form/function.