Add MsbmsBuildLib systems component list generate
In the Hugo/Docsy documentation we have a page of all the various Elixir
Components in the system with the details of their dependencies and
links to their Elixir and database documentation.  Which the amount of
refactoring recently done, it two things occurred to me: 1) this
document could be algorithmically constructed from the Elixir components
themselves and with other information that from our configuration files;
and 2) this would be a good LLM task since it's not in the primary code-
path of the application.

So in this commit we implement the following:

  - Update the MsbmsBuiltLib library to include the automatic generation
    of the system component's list in the documentation.  This includes
    implementation and mix tasks, etc.

  - Create a new documentation standard for Component `README.md` files.
    We now will accept content between the `<!-- MDOC !-->` and
    `<!-- MDESC !-->` designators in the the README.md file to mark the
    boundaries of short descriptive text for the system components list.

Note that this was very much vibe-coded with the LLM doing the vast
majority of the work and documentation.  In fact I think there will be
some changes at some point (naming, etc), but for right now it works
well enough that I can generate that system components list with very
little effort or concern.  I will finish by saying that my review of
this new code has been minimal, looking more at the results of realistic
scenarios running it than actually scrutinizing the code itself; again,
it's outside the critical code path so this doesn't feel as problematic
if it were actual application code.