Adds summary delimiter to README.md files
Our build script can now generate the Systems Component List file based
on the data in the Component `mix.exs` and `README.md` files.  The
README files are used to retrieve the Component Summary paragraphs which
are then pulled into the Component list as summaries there.

To accomplish this the script looks for the origin `<!-- MDOC !-->` tags
and the new `<!-- MDESC !-->` tab being added to the README files in
this commit (largely so anyway).  The text between those tags is the
summary used by the script for the Components list.