Fix NimbleOptions type_doc missing back-ticks
Several places were found which were missing back-ticks in the
NimbleOptions schema definitions.  In some cases these were unbalanced
and in others just missing altogether.

Finally also fixed a documentation reference to a type which was missing
a "t:" prefix and so was causing ExDoc to assume the type reference was
a function reference.