Remove incomplete Components from source tree
This commit removes from the source tree Components which were only ever
partially Completed or which require serious reformulation since the
substantial refactoring that we've done over the past months.

Rather than keep these Components in the source tree until we're ready
to address them, move them out of the tree for more intentional
reintroduction.  None of the code being removed in this commit has been
discarded and is being managed outside of this main source tree until
such time that we should reintroduce them.

Perhaps we'll bring them back under their own feature branches which we
can keep up to date in a more sensible way than just dragging them
along with whatever we happen to be working on.