Fix MsbmsBuildLib System Components doc generation
The imported README.md from each Component was being imported and just
interpolated in as-is.  This meant that all text after the first line
wasn't indented at all relative to the larger System Components List
formatting.

In this commit we add the contextual indentation as well as correct the
in-document anchors which need to match the headings (no "_") but
down-cased.

The overall LLM code quality (this library was pretty well vibe-coded)
is pretty poor and complex, but it's not critical path so we'll just
hack our fixes in and hope we don't have reason to revisit this stuff
often; if we do end up having to really maintain this we'll probably
want to rewrite it ourselves.