Add MscmpSystDb test_support Dialyzer fix
This is actually an LLM recommendation about an error I'm seeing in
CI, but cannot reproduce locally.  I think the LLM is right and that
I've ending up with an environmental ignore missing returns, whereas
CI is running as test environment and is not getting that flag.

If this moves the error, I may have more of these and the LLM analysis
sounds plausible (and probable).