Add MscmpSystPerms.TestSelector
The new Elixir type system checks for protocol implementations and makes
loud warnings if there are none.  Since `MscmpSystPerms` is basically a
library establishing a protocol for other more specialized libraries to
implement, we have no implementation of our own here and thus get the
warning.

To silence these warnings, mostly, we add an implementation of the
protocol to dev/test builds.  We never actually use the implementation,
but its presence is enough to shut the type system up.  Note that the
implementation is entirely LLM generated with human review; it simply
returns empty values of the expected types... so not much to go wrong...
perfect use of LLM.