Remove MscmpSystOptions negative tests on maps
The new Elixir 1.19 type system appears to catch the typing mismatch
early and loudly, so I don't think negative tests are required here; but
more than that, the type system saw the mismatches in the tests and
complained about those... and I don't believe those warnings can be
suppressed.

Remove the tests to get rid of the expected warnings which make it more
likely to miss actually important test failure messages.