Fix MscmpSystAuthn DateTime related unit tests
Elixir 1.19 made a "fix" to `DateTime.diff/2` which removes the
truncation of certain time elements in processing the diff.  Our tests
are on such short timescales that previously truncated times were
essentially the same, but the high precision values are different in
ways that matter to our test... causing them to start to fail.

Explicitly ensure that both times in the test have the same precision by
adopting the least precise time as the standard, truncating the more
precise time.