Fix MscmpSystAuthn MapSet complaint/simplify
Elixir 1.19/OTP28 changes to opacity and MapSets caused typing related
errors to start to show up.

Working with the LLM, the original test seemed convoluted and so was
replaced with something much simpler which also resolves the opacity
complaint.

We may be able to simplify further as I'm not sure we need a MapSet at
all anymore.... but question for another day.