Remove old book version based on Docsy submodule
In preparation for switching to the Jujustu version control system, we
want to simplify our existing repository by getting rid of the Docsy
git submodule installation.  JJ doesn't handle that very well and it's
not even the recommended way to deal with Docsy now either.

In this commit we simply delete the original version of the Book from
the source tree.

I've copied the entire `book` directory out of the tree.  We'll start a
new "Book" instance in the source tree using the current versions of
Hugo and Docsy; naturally, this will be implementing Docsy as a Hugo
module rather than a submodule.  As things look good, we'll slowly
migrate over our original files, fixing any issues along the way.